Python 3 compatibility Update to mamele 4.3.187

+++ setup.py
@@ -16,12 +16,13 @@ from setuptools.command.sdist import sdist as SetuptoolsSdist
 MaxCPUs = 4
 
 # tag to request
-mamele_version = 'v4.2.183'
+mamele_version = 'v4.3.187'
 
 current_directory = os.path.dirname(__file__)
 module_directory = os.path.join(current_directory, 'mamele')
 mamele_versioned = os.path.join(module_directory, 'mamele_%s' % mamele_version)
 mamele_base_directory = os.path.join(module_directory, 'mamele_real')
+mamele_dev_directory = os.path.join(module_directory, 'mamele_dev')
 bindings_directory = os.path.join(mamele_base_directory, 'learning_environment', 'example_agents')
 description_file_directory = os.path.join(mamele_base_directory, 'learning_environment')
 mame_binary_directory = mamele_base_directory
@@ -31,7 +32,7 @@ def module_relative(path):
 
 package_data = [
     module_relative(os.path.join(mamele_base_directory, 'mame64')),
-    module_relative(os.path.join(bindings_directory, 'pythonbinding.so')),
+    module_relative(os.path.join(bindings_directory, 'python%sbinding.so' % sys.version_info.major)),
     module_relative(os.path.join(description_file_directory, 'gameover_description.txt')),
     module_relative(os.path.join(description_file_directory, 'score_description.txt')),
     ]
@@ -49,53 +50,58 @@ class Build(DistutilsBuild):
         # 'mamele' or some URL, and that 'pip install .[mame]' work from gym
         # (https://github.com/openai/gym)
 
-        if not os.path.exists(mamele_versioned):
-            # do a shallow clone
-            git_command = ['git', 'clone', '-b', mamele_version, '--depth', '1', 'https://github.com/alito/mamele.git', mamele_versioned]
+
+        # If there's a mamele_dev directory or link, we assume we are in 
+        # development mode and that the developer knows what they are doing
+        if not os.path.exists(mamele_dev_directory):
+
+            if not os.path.exists(mamele_versioned):
+                # do a shallow clone
+                git_command = ['git', 'clone', '-b', mamele_version, '--depth', '1', 'https://github.com/alito/mamele.git', mamele_versioned]
+                try:
+                    subprocess.check_call(git_command)
+                except subprocess.CalledProcessError as e:
+                    print("Could not do a git clone of the mamele source: %s." % e, file=sys.stderr)
+                    raise
+                except OSError as e:
+                    print("Unable to execute '{}'. You need to have a working 'git' command for this to work".format(" ".join(git_command)), file=sys.stderr)
+                    raise
+
+            if os.path.lexists(mamele_base_directory):
+                if os.path.islink(mamele_base_directory) and not os.path.abspath(os.readlink(mamele_base_directory)) != os.path.abspath(mamele_versioned):
+                    # Assume we are pointing to an outdated version
+                    # Remove and relink
+                    os.remove(mamele_base_directory)
+                    os.symlink(os.path.relpath(mamele_versioned, os.path.dirname(mamele_base_directory)), mamele_base_directory)
+                # if it is not a link, we can assume the user knows better
+            else:
+                os.symlink(os.path.relpath(mamele_versioned, os.path.dirname(mamele_base_directory)), mamele_base_directory)
+
+
+            cores_to_use = min(MaxCPUs, max(1, multiprocessing.cpu_count() - 1))
+
+            print("Compiling MAME, this takes a while", file=sys.stderr)
+            # Compile main MAME
+            cmd = ['make', '-C', mamele_base_directory, '-j', str(cores_to_use)]
             try:
-                subprocess.check_call(git_command)
+                subprocess.check_call(cmd)
             except subprocess.CalledProcessError as e:
-                print("Could not do a git clone of the mamele source: %s." % e, file=sys.stderr)
+                print("Could not build mamele: %s." % e, file=sys.stderr)
                 raise
             except OSError as e:
-                print("Unable to execute '{}'. You need to have a working 'git' command for this to work".format(" ".join(git_command)), file=sys.stderr)
+                print("Unable to execute '{}'. HINT: are you sure `make` is installed?".format(' '.join(cmd)), file=sys.stderr)
                 raise
 
-        if os.path.lexists(mamele_base_directory):
-            if os.path.islink(mamele_base_directory) and not os.path.abspath(os.readlink(mamele_base_directory)) != os.path.abspath(mamele_versioned):
-                # Assume we are pointing to an outdated version
-                # Remove and relink
-                os.remove(mamele_base_directory)
-                os.symlink(os.path.relpath(mamele_versioned, os.path.dirname(mamele_base_directory)), mamele_base_directory)
-            # if it is not a link, we can assume the user knows better
-        else:
-            os.symlink(os.path.relpath(mamele_versioned, os.path.dirname(mamele_base_directory)), mamele_base_directory)
-
-
-        cores_to_use = min(MaxCPUs, max(1, multiprocessing.cpu_count() - 1))
-
-        print("Compiling MAME, this takes a while", file=sys.stderr)
-        # Compile main MAME
-        cmd = ['make', '-C', mamele_base_directory, '-j', str(cores_to_use)]
-        try:
-            subprocess.check_call(cmd)
-        except subprocess.CalledProcessError as e:
-            print("Could not build mamele: %s." % e, file=sys.stderr)
-            raise
-        except OSError as e:
-            print("Unable to execute '{}'. HINT: are you sure `make` is installed?".format(' '.join(cmd)), file=sys.stderr)
-            raise
-
-        # Compile Python bindings
-        cmd = ['make', '-C', bindings_directory , '-j', str(cores_to_use)]
-        try:
-            subprocess.check_call(cmd)
-        except subprocess.CalledProcessError as e:
-            print("Could not build mamele's Python bindings: %s." % e, file=sys.stderr)
-            raise
-        except OSError as e:
-            print("Unable to execute '{}'. HINT: are you sure `make` is installed?\n".format(' '.join(cmd)), file=sys.stderr)
-            raise
+            # Compile Python bindings
+            cmd = ['make', '-C', bindings_directory , '-j', str(cores_to_use)]
+            try:
+                subprocess.check_call(cmd)
+            except subprocess.CalledProcessError as e:
+                print("Could not build mamele's Python bindings: %s." % e, file=sys.stderr)
+                raise
+            except OSError as e:
+                print("Unable to execute '{}'. HINT: are you sure `make` is installed?\n".format(' '.join(cmd)), file=sys.stderr)
+                raise
 
         DistutilsBuild.run(self)
 
@@ -116,7 +122,7 @@ class Sdist(SetuptoolsSdist):
 
 
 setup(name='mamele',
-      version='4.2.183',
+      version='4.3.187',
       description='Python bindings to MAME games',
       long_description='This is a Python wrapper around mamele, a framework for putting computer programs as players of games supported by MAME',
       url='https://github.com/alito/mamele_pippable',
@@ -130,6 +136,8 @@ setup(name='mamele',
         'Intended Audience :: Science/Research',
         'Programming Language :: Python :: 2',
         'Programming Language :: Python :: 2.7',
+        'Programming Language :: Python :: 3',
+        'Programming Language :: Python :: 3.5',
         'License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)',
         'License :: OSI Approved :: BSD License',
         'Operating System :: POSIX :: Linux',
